リファクタリング
git-svn-id: https://www.seasar.org/svn/sandbox/s2chronos/trunk@131 5fbd9426-1b17-45cb-9c0a-1918607842a3

+++ s2chronos/src/main/java/org/seasar/chronos/store/dao/CronExpressionDao.java
@@ -12,7 +12,7 @@ public interface CronExpressionDao {
 
 	public int delete(CronExpressionEntity entity);
 
-	public CronExpressionEntity selectById(Long id);
+	public CronExpressionEntity selectById(int id);
 
 	public List<CronExpressionEntity> selectAll();
