+++ s2chronos/src/main/java/org/seasar/chronos/store/dao/TriggerDao.java
@@ -12,7 +12,7 @@ public interface TriggerDao {
 
 	public int delete(TriggerEntity entity);
 
-	public TriggerEntity selectById(Long id);
+	public TriggerEntity selectById(int id);
 
 	public List<TriggerEntity> selectAll();
